Finish fitness function; cost = profit * - 1

+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optimiser/src/main/java/com/mmxlabs/scheduler/optimiser/fitness/components/allocation/ICargoAllocationEditor.java
@@ -15,5 +15,5 @@ import com.mmxlabs.scheduler.optimiser.components.IPortSlot;
  *
  */
 public interface ICargoAllocationEditor<T> extends ICargoAllocationProvider<T> {
-	public void addCargoAllocationLimit(final Set<IPortSlot> slots, final int maximumQuantity);
+	public void addCargoAllocationLimit(final Set<IPortSlot> slots, final long maximumQuantity);
 }
